Move logic to CardView
