+++ forge-gui/src/main/java/forge/gamemodes/match/HostedMatch.java
@@ -18,6 +18,7 @@ import com.google.common.collect.Maps;
 import com.google.common.eventbus.Subscribe;
 
 import forge.LobbyPlayer;
+import forge.StaticData;
 import forge.game.Game;
 import forge.game.GameRules;
 import forge.game.GameType;
@@ -156,7 +157,7 @@ public class HostedMatch {
 
         game = match.createGame();
 
-        game.getView().setSourceImageForClone(FModel.getPreferences().getPrefBoolean(FPref.UI_CLONE_MODE_SOURCE));
+        StaticData.instance().setSourceImageForClone(FModel.getPreferences().getPrefBoolean(FPref.UI_CLONE_MODE_SOURCE));
 
         if (game.getRules().getGameType() == GameType.Quest) {
             final QuestController qc = FModel.getQuest();
